Add consumable effect activation feature
Right-clicking on a consumable in the consumable panel now toggles its effect. Consumable effects of activated consumables affecting the ship stats (temporarily or permanent) are included in the effective ship stats calculation and are shown in the ShipStatsControl.

Due to parallelization during the consumable viewmodel generation, the AircraftCache of the AppData class was changed into a ConcurrentDictionary because this data is required by some consumables.

Relates to #43