Fix captain skill buttons being disabled after changing captains
If captain skills are selected and the captain is changed, all skills which require more skill points than are currently free become disabled and cannot be toggled anymore.
In case of a 21 point build this results in captain skills becoming fixed completely. Otherwise, skills can be changed after changing one of the still enabled skills.