Fix reordering of selected skills when more than one skill is unordered.
Due to creating a copy of the initial, unordered list and using this copy to determine the indexes for insertion of sorted elements, the elements were inserted in wrong positions when more than a single element had to be moved. Now the FindIndex method has been added to CustomObservableCollection and the skill list is no longer copied.